% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_mult_cube_vec}
\alias{arma_mult_cube_vec}
\title{Cube multiply over first dimension}
\usage{
arma_mult_cube_vec(cub, v)
}
\arguments{
\item{cub}{A cube (3D array)}

\item{v}{A vector}
}
\value{
Multiplication over first dimension of cub time v
}
\description{
Cube multiply over first dimension
}
\examples{
corr_gauss_matrix_sym_armaC(matrix(c(1,0,0,1),2,2),c(1,1))

x3 <- matrix(runif(1e3*6), ncol=6)
th <- c(.3,3.3)
t3 <- corr_gauss_matrix_symC(x3, th)
t4 <- corr_gauss_matrix_sym_armaC(x3, th)
identical(t3, t4)
# microbenchmark::microbenchmark(corr_gauss_matrix_symC(x3, th),
#                     corr_gauss_matrix_sym_armaC(x3, th), times=50)
}
