% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corr_cubic_matrix_symC}
\alias{corr_cubic_matrix_symC}
\title{Correlation Cubic matrix in C (symmetric)}
\usage{
corr_cubic_matrix_symC(x, theta)
}
\arguments{
\item{x}{Matrix x}

\item{theta}{Theta vector}
}
\value{
Correlation matrix
}
\description{
Correlation Cubic matrix in C (symmetric)
}
\examples{
corr_cubic_matrix_symC(matrix(c(1,0,0,1),2,2),c(1,1))
}
