% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corr_gauss_dCdX}
\alias{corr_gauss_dCdX}
\title{Correlation Gaussian matrix gradient in C using Armadillo}
\usage{
corr_gauss_dCdX(XX, X, theta, s2)
}
\arguments{
\item{XX}{Matrix XX to get gradient for}

\item{X}{Matrix X GP was fit to}

\item{theta}{Theta vector}

\item{s2}{Variance parameter}
}
\value{
3-dim array of correlation derivative
}
\description{
Correlation Gaussian matrix gradient in C using Armadillo
}
\examples{
# corr_gauss_dCdX(matrix(c(1,0,0,1),2,2),c(1,1))
}
