% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr_gauss_matrix}
\alias{corr_gauss_matrix}
\title{Gaussian correlation}
\usage{
corr_gauss_matrix(x, x2 = NULL, theta)
}
\arguments{
\item{x}{First data matrix}

\item{x2}{Second data matrix}

\item{theta}{Correlation parameter}
}
\value{
Correlation matrix
}
\description{
Gaussian correlation
}
\examples{
corr_gauss_matrix(matrix(1:10,ncol=1), matrix(6:15,ncol=1), 1e-2/(1:10))
}
