% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corr_gauss_matrixC}
\alias{corr_gauss_matrixC}
\title{Correlation Gaussian matrix in C using Rcpp}
\usage{
corr_gauss_matrixC(x, y, theta)
}
\arguments{
\item{x}{Matrix x}

\item{y}{Matrix y, must have same number of columns as x}

\item{theta}{Theta vector}
}
\value{
Correlation matrix
}
\description{
Correlation Gaussian matrix in C using Rcpp
}
\examples{
corr_gauss_matrixC(matrix(c(1,0,0,1),2,2), matrix(c(1,0,1,1),2,2), c(1,1))
}
