% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corr_gauss_matrixCpar}
\alias{corr_gauss_matrixCpar}
\title{Correlation Gaussian matrix in C using RcppParallel}
\usage{
corr_gauss_matrixCpar(mat1, mat2, theta)
}
\arguments{
\item{theta}{Theta vector}

\item{x}{Matrix x}

\item{y}{Matrix y, must have same number of columns as x}
}
\value{
Correlation matrix
}
\description{
Faster than nonparallel version for D < 12 and > 20 rows
}
\examples{
corr_gauss_matrixC(matrix(c(1,0,0,1),2,2), matrix(c(1,0,1,1),2,2), c(1,1))
}
