% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corr_latentfactor_matrixmatrixC}
\alias{corr_latentfactor_matrixmatrixC}
\title{Correlation Latent factor  matrix in C (symmetric)}
\usage{
corr_latentfactor_matrixmatrixC(x, y, theta, xindex, latentdim, offdiagequal)
}
\arguments{
\item{x}{Matrix x}

\item{y}{Matrix y}

\item{theta}{Theta vector}

\item{xindex}{Index to use}

\item{latentdim}{Number of latent dimensions}

\item{offdiagequal}{What to set off-diagonal values with matching values to.}
}
\value{
Correlation matrix
}
\description{
Correlation Latent factor  matrix in C (symmetric)
}
\examples{
corr_latentfactor_matrixmatrixC(matrix(c(1,.5, 2,1.6, 1,0),ncol=2,byrow=TRUE),
                                matrix(c(2,1.6, 1,0),ncol=2,byrow=TRUE),
                                c(1.5,1.8), 1, 1, 1-1e-6)
corr_latentfactor_matrixmatrixC(matrix(c(0,0,0,1,0,0,0,2,0,0,0,3,0,0,0,4),
                                  ncol=4, byrow=TRUE),
                                matrix(c(0,0,0,2,0,0,0,4,0,0,0,1),
                                  ncol=4, byrow=TRUE),
  c(0.101, -0.714, 0.114, -0.755, 0.117, -0.76, 0.116, -0.752),
  4, 2, 1-1e-6) * 6.85
}
