% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGO.R
\docType{class}
\name{EGO}
\alias{EGO}
\title{EGO R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
EGO
}
\value{
Object of \code{\link{R6Class}} with methods for running EGO.
}
\description{
EGO R6 class
}
\examples{
e1 <- EGO$new(func=sin, n0=10, n=10, d=1)
}
\keyword{Bayesian}
\keyword{optimization}
\keyword{optimization,}
