% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Exponential.R
\docType{class}
\name{Exponential}
\alias{Exponential}
\title{Exponential Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
Exponential
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Exponential Kernel R6 class
}
\examples{
k1 <- Exponential$new(beta=0)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
