% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPr_Gauss_par.R
\docType{class}
\name{GauPr_Gauss_par}
\alias{GauPr_Gauss_par}
\title{Corr Gauss GP using inherited optim}
\format{\code{\link{R6Class}} object.}
\usage{
GauPr_Gauss_par
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Corr Gauss GP using inherited optim
}
\examples{
n <- 12
x <- matrix(seq(0,1,length.out = n), ncol=1)
y <- sin(2*pi*x) + rnorm(n,0,1e-1)
gp <- GauPro$new(X=x, Z=y, parallel=FALSE)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}

