% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPro_selector.R
\name{GauPro}
\alias{GauPro}
\title{GauPro_selector}
\usage{
GauPro(..., type = "Gauss")
}
\arguments{
\item{...}{Pass on}

\item{type}{Type of Gaussian process, or the kind of correlation function.}
}
\value{
A GauPro object
}
\description{
GauPro_selector
}
\examples{

}

