% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_beta.R
\docType{class}
\name{GauPro_kernel_beta}
\alias{GauPro_kernel_beta}
\title{Beta Kernel R6 class}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
This is the base structure for a kernel that uses beta = log10(theta)
for the lengthscale parameter.
It standardizes the params because they all use the same underlying structure.
Kernels that inherit this only need to implement kone and dC_dparams.
}
\examples{
#k1 <- Matern52$new(beta=0)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
\section{Super class}{
\code{\link[GauPro:GauPro_kernel]{GauPro::GauPro_kernel}} -> \code{GauPro_kernel_beta}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GauPro_kernel_beta$new()}}
\item \href{#method-k}{\code{GauPro_kernel_beta$k()}}
\item \href{#method-kone}{\code{GauPro_kernel_beta$kone()}}
\item \href{#method-param_optim_start}{\code{GauPro_kernel_beta$param_optim_start()}}
\item \href{#method-param_optim_start0}{\code{GauPro_kernel_beta$param_optim_start0()}}
\item \href{#method-param_optim_lower}{\code{GauPro_kernel_beta$param_optim_lower()}}
\item \href{#method-param_optim_upper}{\code{GauPro_kernel_beta$param_optim_upper()}}
\item \href{#method-set_params_from_optim}{\code{GauPro_kernel_beta$set_params_from_optim()}}
\item \href{#method-C_dC_dparams}{\code{GauPro_kernel_beta$C_dC_dparams()}}
\item \href{#method-s2_from_params}{\code{GauPro_kernel_beta$s2_from_params()}}
\item \href{#method-clone}{\code{GauPro_kernel_beta$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$new(
  beta,
  s2 = 1,
  D,
  beta_lower = -8,
  beta_upper = 6,
  beta_est = TRUE,
  s2_lower = 1e-08,
  s2_upper = 1e+08,
  s2_est = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-k"></a>}}
\if{latex}{\out{\hypertarget{method-k}{}}}
\subsection{Method \code{k()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$k(
  x,
  y = NULL,
  beta = self$beta,
  s2 = self$s2,
  params = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kone"></a>}}
\if{latex}{\out{\hypertarget{method-kone}{}}}
\subsection{Method \code{kone()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$kone(x, y, beta, theta, s2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-param_optim_start"></a>}}
\if{latex}{\out{\hypertarget{method-param_optim_start}{}}}
\subsection{Method \code{param_optim_start()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$param_optim_start(
  jitter = F,
  y,
  beta_est = self$beta_est,
  s2_est = self$s2_est
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-param_optim_start0"></a>}}
\if{latex}{\out{\hypertarget{method-param_optim_start0}{}}}
\subsection{Method \code{param_optim_start0()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$param_optim_start0(
  jitter = F,
  y,
  beta_est = self$beta_est,
  s2_est = self$s2_est
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-param_optim_lower"></a>}}
\if{latex}{\out{\hypertarget{method-param_optim_lower}{}}}
\subsection{Method \code{param_optim_lower()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$param_optim_lower(
  beta_est = self$beta_est,
  s2_est = self$s2_est
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-param_optim_upper"></a>}}
\if{latex}{\out{\hypertarget{method-param_optim_upper}{}}}
\subsection{Method \code{param_optim_upper()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$param_optim_upper(
  beta_est = self$beta_est,
  s2_est = self$s2_est
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_params_from_optim"></a>}}
\if{latex}{\out{\hypertarget{method-set_params_from_optim}{}}}
\subsection{Method \code{set_params_from_optim()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$set_params_from_optim(
  optim_out,
  beta_est = self$beta_est,
  s2_est = self$s2_est
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-C_dC_dparams"></a>}}
\if{latex}{\out{\hypertarget{method-C_dC_dparams}{}}}
\subsection{Method \code{C_dC_dparams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$C_dC_dparams(params = NULL, X, nug)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-s2_from_params"></a>}}
\if{latex}{\out{\hypertarget{method-s2_from_params}{}}}
\subsection{Method \code{s2_from_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$s2_from_params(params, s2_est = self$s2_est)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GauPro_kernel_beta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
