% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_base.R
\docType{class}
\name{GauPro_trend}
\alias{GauPro_trend}
\title{Trend R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
GauPro_trend
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Trend R6 class
}
\examples{
k <- GauPro_trend$new()
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
