% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Gaussian.R
\docType{class}
\name{Gaussian_bad}
\alias{Gaussian_bad}
\title{Gaussian Kernel R6 class- bad since params aren't on log scale}
\format{\code{\link{R6Class}} object.}
\usage{
Gaussian_bad
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Gaussian Kernel R6 class- bad since params aren't on log scale
}
\examples{
k1 <- Gaussian_bad$new(theta=1)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
