% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Gaussian_beta.R
\docType{class}
\name{Gaussian_beta}
\alias{Gaussian_beta}
\title{Gaussian Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
Gaussian_beta
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Gaussian Kernel R6 class
}
\examples{
k1 <- Gaussian_beta$new(beta=0)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
