% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gaussian_devianceC}
\alias{Gaussian_devianceC}
\title{Calculate the Gaussian deviance in C}
\usage{
Gaussian_devianceC(theta, nug, X, Z)
}
\arguments{
\item{theta}{Theta vector}

\item{nug}{Nugget}

\item{X}{Matrix X}

\item{Z}{Matrix Z}
}
\value{
Correlation matrix
}
\description{
Calculate the Gaussian deviance in C
}
\examples{
Gaussian_devianceC(c(1,1), 1e-8, matrix(c(1,0,0,1),2,2), matrix(c(1,0),2,1))
}
