% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gaussian_hessianCC}
\alias{Gaussian_hessianCC}
\title{Gaussian hessian in C}
\usage{
Gaussian_hessianCC(XX, X, Z, Kinv, mu_hat, theta)
}
\arguments{
\item{XX}{point to find Hessian at}

\item{X}{matrix of data points}

\item{Z}{matrix of output}

\item{Kinv}{inverse of correlation matrix}

\item{mu_hat}{mean estimate}

\item{theta}{correlation parameters}
}
\value{
Hessian matrix
}
\description{
Gaussian hessian in C
}
