% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Gaussian_l.R
\docType{class}
\name{Gaussian_l}
\alias{Gaussian_l}
\title{Gaussian Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
Gaussian_l
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Gaussian Kernel R6 class
}
\examples{
k1 <- Gaussian_l$new(l=1)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
