% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradfuncarray}
\alias{gradfuncarray}
\title{Calculate gradfunc in optimization to speed up.
NEEDS TO APERM dC_dparams
Doesn't need to be exported, should only be useful in functions.}
\usage{
gradfuncarray(dC_dparams, Cinv, Cinv_yminusmu)
}
\arguments{
\item{dC_dparams}{Derivative matrix for covariance function wrt kernel parameters}

\item{Cinv}{Inverse of covariance matrix}

\item{Cinv_yminusmu}{Vector that is the inverse of C times y minus the mean.}
}
\value{
Vector, one value for each parameter
}
\description{
Calculate gradfunc in optimization to speed up.
NEEDS TO APERM dC_dparams
Doesn't need to be exported, should only be useful in functions.
}
\examples{
# corr_gauss_dCdX(matrix(c(1,0,0,1),2,2),c(1,1))
}
