% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradfuncarray.R
\name{gradfuncarrayR}
\alias{gradfuncarrayR}
\title{Calculate gradfunc in optimization to speed up.
NEEDS TO APERM dC_dparams
Doesn't need to be exported, should only be useful in functions.}
\usage{
gradfuncarrayR(dC_dparams, Cinv, Cinv_yminusmu)
}
\arguments{
\item{dC_dparams}{Derivative matrix for covariance function wrt kernel parameters}

\item{Cinv}{Inverse of covariance matrix}

\item{Cinv_yminusmu}{Vector that is the inverse of C times y minus the mean.}
}
\value{
Vector, one value for each parameter
}
\description{
Calculate gradfunc in optimization to speed up.
NEEDS TO APERM dC_dparams
Doesn't need to be exported, should only be useful in functions.
}
\examples{
a1 <- array(dim=c(2,4,4), data=rnorm(32))
a2 <- matrix(rnorm(16),4,4)
a3 <- rnorm(4)
#gradfuncarray(a1, a2, a3)
#gradfuncarrayR(a1, a2, a3)
}
