% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernel_orderedFactor_dC}
\alias{kernel_orderedFactor_dC}
\title{Derivative of covariance matrix of X with respect to kernel
parameters for the Ordered Factor Kernel}
\usage{
kernel_orderedFactor_dC(
  x,
  pf,
  C_nonug,
  s2_est,
  p_est,
  lenparams_D,
  s2_nug,
  xindex,
  nlevels,
  s2
)
}
\arguments{
\item{x}{Matrix x}

\item{pf}{pf vector}

\item{C_nonug}{cov mat without nugget}

\item{s2_est}{whether s2 is being estimated}

\item{p_est}{Whether theta/beta is being estimated}

\item{lenparams_D}{Number of parameters the derivative is being calculated for}

\item{s2_nug}{s2 times the nug}

\item{xindex}{Which column of x is the indexing variable}

\item{nlevels}{Number of levels}

\item{s2}{Value of s2}
}
\value{
Correlation matrix
}
\description{
Derivative of covariance matrix of X with respect to kernel
parameters for the Ordered Factor Kernel
}
