% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_product.R
\docType{class}
\name{kernel_product}
\alias{kernel_product}
\title{Gaussian Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
kernel_product
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Gaussian Kernel R6 class
}
\examples{
k1 <- Exponential$new(beta=1)
k2 <- Matern32$new(beta=2)
k <- k1 + k2
k$k(matrix(c(2,1), ncol=1))
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
