% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_sum.R
\docType{class}
\name{kernel_sum}
\alias{kernel_sum}
\title{Gaussian Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
kernel_sum
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Gaussian Kernel R6 class
}
\examples{

}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
