% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Matern32.R
\docType{class}
\name{Matern32}
\alias{Matern32}
\title{Matern 3/2 Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
Matern32
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Matern 3/2 Kernel R6 class
}
\examples{
k1 <- Matern32$new(beta=0)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
