% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Matern52.R
\docType{class}
\name{Matern52}
\alias{Matern52}
\title{Matern 5/2 Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
Matern52
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Matern 5/2 Kernel R6 class
}
\examples{
k1 <- Matern52$new(beta=0)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
