% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_periodic.R
\docType{class}
\name{Periodic}
\alias{Periodic}
\title{Periodic Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
Periodic
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Periodic Kernel R6 class
}
\examples{
k1 <- Periodic$new(p=1, alpha=1)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
