% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPro_S3.R
\name{plot.GauPro}
\alias{plot.GauPro}
\title{Plot for class GauPro}
\usage{
\method{plot}{GauPro}(x, ...)
}
\arguments{
\item{x}{Object of class GauPro}

\item{...}{Additional parameters}
}
\value{
Nothing
}
\description{
Plot for class GauPro
}
\examples{
n <- 12
x <- matrix(seq(0,1,length.out = n), ncol=1)
y <- sin(2*pi*x) + rnorm(n,0,1e-1)
gp <- GauPro(X=x, Z=y, parallel=FALSE)
if (requireNamespace("MASS", quietly = TRUE)) {
  plot(gp)
}

}
