% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPro_S3.R
\name{+.GauPro_kernel}
\alias{+.GauPro_kernel}
\title{Kernel sum}
\usage{
\method{+}{GauPro_kernel}(k1, k2)
}
\arguments{
\item{k1}{First kernel}

\item{k2}{Second kernel}
}
\value{
Kernel which is sum of two kernels
}
\description{
Kernel sum
}
\examples{
k1 <- Exponential$new(beta=1)
k2 <- Matern32$new(beta=0)
k <- k1 + k2
k$k(matrix(c(2,1), ncol=1))
}
