% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_PowerExp.R
\docType{class}
\name{PowerExp}
\alias{PowerExp}
\title{Power Exponential Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
PowerExp
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Power Exponential Kernel R6 class
}
\examples{
k1 <- PowerExp$new(beta=0, alpha=0)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
