% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPro_base.R
\name{predict.GauPro_base}
\alias{predict.GauPro_base}
\title{Predict for class GauPro_base}
\usage{
\method{predict}{GauPro_base}(object, XX, se.fit = F, covmat = F, split_speed = T, ...)
}
\arguments{
\item{object}{Object of class GauPro_base}

\item{XX}{Points to predict at}

\item{se.fit}{Should the se be returned?}

\item{covmat}{Should the covariance matrix be returned?}

\item{split_speed}{Should the predictions be split up for speed}

\item{...}{Additional parameters}
}
\value{
Prediction from object at XX
}
\description{
Predict mean and se for given matrix
}
