% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPro_S3.R
\name{print.summary.GauPro}
\alias{print.summary.GauPro}
\title{Print summary.GauPro}
\usage{
\method{print}{summary.GauPro}(x, ...)
}
\arguments{
\item{x}{summary.GauPro object}

\item{...}{Additional args}
}
\value{
prints, returns invisible object
}
\description{
Print summary.GauPro
}
