% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt_matrix.R
\name{sqrt_matrix}
\alias{sqrt_matrix}
\title{Find the square root of a matrix}
\usage{
sqrt_matrix(mat, symmetric)
}
\arguments{
\item{mat}{Matrix to find square root matrix of}

\item{symmetric}{Is it symmetric? Passed to eigen.}
}
\value{
Square root of mat
}
\description{
Same thing as 'expm::sqrtm', but faster.
}
\examples{
mat <- matrix(c(1,.1,.1,1), 2, 2)
smat <- sqrt_matrix(mat=mat, symmetric=TRUE)
smat \%*\% smat
}
