% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPro_S3.R
\name{summary.GauPro}
\alias{summary.GauPro}
\title{Summary for GauPro object}
\usage{
\method{summary}{GauPro}(object, ...)
}
\arguments{
\item{object}{GauPro R6 object}

\item{...}{Additional arguments passed to summary}
}
\value{
Summary
}
\description{
Summary for GauPro object
}
