% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_0.R
\docType{class}
\name{trend_0}
\alias{trend_0}
\title{Trend R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
trend_0
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Trend R6 class
}
\examples{
t1 <- trend_0$new()
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
