% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_constant.R
\docType{class}
\name{trend_c}
\alias{trend_c}
\title{Trend R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
trend_c
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Trend R6 class
}
\examples{
t1 <- trend_c$new()
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
