% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Triangle.R
\docType{class}
\name{Triangle}
\alias{Triangle}
\title{Triangle Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
Triangle
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Triangle Kernel R6 class
}
\examples{
k1 <- Triangle$new(beta=0)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
