% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_White.R
\docType{class}
\name{White}
\alias{White}
\title{White noise Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
White
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
White noise Kernel R6 class
}
\examples{
k1 <- White$new(s2=1e-8)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
