% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corr_gauss_matrix_sym_armaC}
\alias{corr_gauss_matrix_sym_armaC}
\title{Correlation Gaussian matrix in C using Armadillo (symmetric)}
\usage{
corr_gauss_matrix_sym_armaC(x, theta)
}
\arguments{
\item{x}{Matrix x}

\item{theta}{Theta vector}
}
\value{
Correlation matrix
}
\description{
About 30% faster than Rcpp version.
}
\examples{
corr_gauss_matrix_sym_armaC(matrix(c(1,0,0,1),2,2),c(1,1))

x3 <- matrix(runif(1e3*6), ncol=6)
th <- runif(6)
t3 <- corr_gauss_matrix_symC(x3, th)
t4 <- corr_gauss_matrix_sym_armaC(x3, th)
identical(t3, t4)
# microbenchmark::microbenchmark(corr_gauss_matrix_symC(x3, th),
#                     corr_gauss_matrix_sym_armaC(x3, th), times=50)
}
