% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPro_selector.R
\name{GauPro}
\alias{GauPro}
\title{GauPro_selector}
\usage{
GauPro(..., type = "Gauss")
}
\arguments{
\item{...}{Pass on}

\item{type}{Type of Gaussian process, or the kind of correlation function.}
}
\value{
A GauPro object
}
\description{
GauPro_selector
}
\examples{
n <- 12
x <- matrix(seq(0,1,length.out = n), ncol=1)
#y <- sin(2*pi*x) + rnorm(n,0,1e-1)
y <- (2*x) \%\%1
gp <- GauPro(X=x, Z=y, parallel=FALSE)
}
