% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauPro_Gauss_LOO.R
\docType{class}
\name{GauPro_Gauss_LOO}
\alias{GauPro_Gauss_LOO}
\title{Corr Gauss GP using inherited optim}
\format{\code{\link{R6Class}} object.}
\usage{
GauPro_Gauss_LOO
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Corr Gauss GP using inherited optim
}
\examples{
n <- 12
x <- matrix(seq(0,1,length.out = n), ncol=1)
y <- sin(2*pi*x) + rnorm(n,0,1e-1)
gp <- GauPro_Gauss_LOO$new(X=x, Z=y, parallel=FALSE)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
