% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_base.R
\docType{class}
\name{GauPro_kernel}
\alias{GauPro_kernel}
\title{Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
GauPro_kernel
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Kernel R6 class
}
\examples{
#k <- GauPro_kernel$new()
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
