% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_beta.R
\docType{class}
\name{GauPro_kernel_beta}
\alias{GauPro_kernel_beta}
\title{Beta Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
GauPro_kernel_beta
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
This is the base structure for a kernel that uses beta = log10(theta)
for the lengthscale parameter.
It standardizes the params because they all use the same underlying structure.
Kernels that inherit this only need to implement kone and dC_dparams.
}
\examples{
#k1 <- Matern52$new(beta=0)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
