% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_RatQuad.R
\docType{class}
\name{RatQuad}
\alias{RatQuad}
\title{Rational Quadratic Kernel R6 class}
\format{\code{\link{R6Class}} object.}
\usage{
RatQuad
}
\value{
Object of \code{\link{R6Class}} with methods for fitting GP model.
}
\description{
Rational Quadratic Kernel R6 class
}
\examples{
k1 <- RatQuad$new(beta=0, alpha=0)
}
\keyword{Gaussian}
\keyword{data,}
\keyword{kriging,}
\keyword{process,}
\keyword{regression}
